[ml] migrate file_data_visualizer/analyze_file to file_upload plugin (#94259)
* [ml] migrate file_data_visualizer/analyze_file to file_upload plugin

* tslint

* give analyze route access to ml user
