+++ x-pack/plugins/ml/public/application/datavisualizer/file_based/components/import_settings/advanced.tsx
@@ -20,7 +20,7 @@ import {
 
 import { CombinedField, CombinedFieldsForm } from '../combined_fields';
 import { MLJobEditor, ML_EDITOR_MODE } from '../../../../jobs/jobs_list/components/ml_job_editor';
-import { FindFileStructureResponse } from '../../../../../../common/types/file_datavisualizer';
+import { FindFileStructureResponse } from '../../../../../../../file_upload/common';
 const EDITOR_HEIGHT = '300px';
 
 interface Props {
