+++ x-pack/plugins/ml/public/application/datavisualizer/file_based/components/results_view/results_view.tsx
@@ -20,7 +20,7 @@ import {
   EuiFlexGroup,
   EuiFlexItem,
 } from '@elastic/eui';
-import { FindFileStructureResponse } from '../../../../../../common/types/file_datavisualizer';
+import { FindFileStructureResponse } from '../../../../../../../file_upload/common';
 
 import { FileContents } from '../file_contents';
 import { AnalysisSummary } from '../analysis_summary';
